+++ python/tests/smoke_test.py
@@ -1,3 +1,0 @@
-import bedrock_bio
-
-assert bedrock_bio.__version__